Extract tree record codec
